requirements.txt
datasets==2.13.1
ipython==8.14.0
keras==2.13.1
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
tensorflow==2.13.0
tensorflow_hub==0.13.0
torch==2.0.1
transformers==4.30.2
django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.2.0
psycopg==3.1.10
celery==5.3.1
